--- src/pages/searchModal/index.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko-KR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>검색 모달</title>
-    <link rel="stylesheet" href="/src/styles/main.scss">
-    <script src="/src/pages/searchModal/searchModal.js" type="module"></script>
-</head>
-<body>
-    <header>
-        <button id="searchBtn">Search</button>
-        <div id="searchModal" class="modal">
-            <div class="modal-content">
-                <div class="input-wrapper">
-                    <label for="searchInput" class="sr-only">컨텐츠 검색창</label>
-                    <input type="text" id="searchInput" placeholder="검색">
-                    <button id="searchSubmitBtn" type="submit"></button>
-                </div>
-                <div class="section-wrapper">
-                    <section class="search-history">
-                    <h1 class="search-history-title">최근 검색어</h1>
-                    <ol class="search-history-list">
-                        <li class="search-history-item">검색 내역이 없습니다.</li>
-                    </ol>
-                    </section>
-                    <section class="popular-searches">
-                        <h1 class="popular-searches-title">실시간 인기 검색어</h1>
-                        <ol class="popular-searches-list">
-                            <li class="popular-searches-item"><a href="#">재벌집 막내아들</a></li>
-                            <li class="popular-searches-item"><a href="#">삼위일체</a></li>
-                            <li class="popular-searches-item"><a href="#">트리니티</a></li>
-                            <li class="popular-searches-item"><a href="#">삼위일체</a></li>
-                            <li class="popular-searches-item"><a href="#">트리니티</a></li>
-                            <li class="popular-searches-item"><a href="#">삼위일체</a></li>
-                            <li class="popular-searches-item"><a href="#">트리니티</a></li>
-                            <li class="popular-searches-item"><a href="#">삼위일체</a></li>
-                            <li class="popular-searches-item"><a href="#">트리니티</a></li>
-                            <li class="popular-searches-item"><a href="#">삼위일체</a></li>
-                        </ol>
-                        <p class="popular-searches-date">2022.12.24 오전 03.04 기준</p>
-                    </section>
-                </div>
-            </div>
-        </div>
-    </header>
-</body>
-</html>
